requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
matplotlib>=3.7.0
pandas>=2.0.0
flask>=3.0.0
plotly>=5.18.0
anthropic>=0.19.0
pillow>=9.0.0